Split the UIColor extension into a different file
